+++ NEWS.md
@@ -2,6 +2,7 @@
 
 # New in version 1.2
 
+* RestructuredText Feature, see <doc/rst.rst>
 * Taxonomies now need to be explicitly listed in settings as a `TAXONOMIES`
   list of taxonomy names. staticsite prints a warning if a `.taxonomy` file is
   found that is not listed in `TAXONOMIES`.